list_service/requirements.txt
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.1
Deprecated==1.2.14
googleapis-common-protos==1.61.0
grpcio==1.48.2
grpcio-tools==1.48.2
idna==3.4
importlib-metadata==6.8.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-zipkin==1.20.0
opentelemetry-exporter-zipkin-json==1.20.0
opentelemetry-exporter-zipkin-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-grpc==0.41b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
prometheus-client==0.17.1
protobuf==3.20.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.15.0
zipp==3.17.0
